# Revamp dejafu-tests