--- a/dejafu-tests/Setup.hs
+++ b/dejafu-tests/Setup.hs
@@ -1,2 +1,2 @@
-import Distribution.Simple
+import           Distribution.Simple
 main = defaultMain