--- a/dejafu-tests/dejafu-tests.cabal
+++ b/dejafu-tests/dejafu-tests.cabal
@@ -16,10 +16,8 @@ build-type:          Simple
 -- extra-source-files:  
 cabal-version:       >=1.10
 
-executable dejafu-tests
-  main-is:             Main.hs
-
-  other-modules:       Cases
+library
+  exposed-modules:     Cases
                      , Cases.Async
                      , Cases.SingleThreaded
                      , Cases.MultiThreaded
@@ -41,7 +39,6 @@ executable dejafu-tests
 
                      , QSemN
 
-  -- other-extensions:    
   build-depends:       base
                      , concurrency
                      , containers
@@ -57,6 +54,15 @@ executable dejafu-tests
                      , test-framework-quickcheck2
   if impl(ghc < 8.0.1)
     build-depends: transformers
-  -- hs-source-dirs:      
+  hs-source-dirs:      lib
   default-language:    Haskell2010
-  ghc-options: -threaded -rtsopts
+
+executable dejafu-tests
+  main-is:             Main.hs
+
+  build-depends:       base
+                     , dejafu-tests
+                     , test-framework
+  hs-source-dirs:      exe
+  default-language:    Haskell2010
+  ghc-options:         -threaded -rtsopts