--- a/dejafu-tests/dejafu-tests.cabal
+++ b/dejafu-tests/dejafu-tests.cabal
@@ -46,7 +46,6 @@ library
 
   build-depends:       base
                      , abstract-deque
-                     , abstract-par
                      , concurrency
                      , containers
                      , deepseq
@@ -67,6 +66,7 @@ library
     build-depends: transformers
   hs-source-dirs:      lib
   default-language:    Haskell2010
+  ghc-options:         -Wall
 
 executable dejafu-tests
   main-is:             Main.hs
@@ -77,4 +77,4 @@ executable dejafu-tests
                      , tasty-quickcheck
   hs-source-dirs:      exe
   default-language:    Haskell2010
-  ghc-options:         -threaded -rtsopts
+  ghc-options:         -Wall -threaded -rtsopts