--- a/dejafu-tests/dejafu-tests.cabal
+++ b/dejafu-tests/dejafu-tests.cabal
@@ -53,12 +53,11 @@ library
                      , hedgehog
                      , mtl
                      , mwc-random
-                     , QuickCheck
                      , random
                      , tasty
+                     , tasty-expected-failure
                      , tasty-dejafu
                      , tasty-hedgehog
-                     , tasty-quickcheck
                      , vector
   if impl(ghc < 8.0.1)
     build-depends: transformers
@@ -72,7 +71,6 @@ executable dejafu-tests
   build-depends:       base
                      , dejafu-tests
                      , tasty
-                     , tasty-quickcheck
   hs-source-dirs:      exe
   default-language:    Haskell2010
   ghc-options:         -Wall -threaded -rtsopts