--- a/dejafu-tests/exe/Main.hs
+++ b/dejafu-tests/exe/Main.hs
@@ -1,6 +1,7 @@
 module Main where
 
 import qualified Test.Tasty            as T
+import qualified Test.Tasty.Options    as T
 import qualified Test.Tasty.QuickCheck as T
 
 import qualified Examples              as E
@@ -8,7 +9,10 @@ import qualified Integration           as I
 import qualified Unit                  as U
 
 main :: IO ()
-main = T.defaultMain $ T.adjustOption reduceQCTests tests
+main =
+  let ingredients = T.includingOptions options : T.defaultIngredients
+      runner = T.defaultMainWithIngredients ingredients
+  in runner (T.adjustOption reduceQCTests tests)
 
 tests :: T.TestTree
 tests = T.testGroup "Tests"
@@ -16,6 +20,9 @@ tests = T.testGroup "Tests"
   , T.testGroup "Integration" I.tests
   , T.testGroup "Examples" E.tests
   ]
+
+options :: [T.OptionDescription]
+options = U.options ++ I.options ++ E.options
 
 -- | Reduce the default number of quickcheck runs.
 reduceQCTests :: T.QuickCheckTests -> T.QuickCheckTests