… .hlint.yaml
… .travis.yml
… .travis/lts-9.0.yaml
… .weeder.yaml
… dejafu-tests/lib/Cases/Regressions.hs
… dejafu-tests/lib/Cases/SingleThreaded.hs
… dejafu-tests/lib/Common.hs
… dejafu-tests/lib/Examples.hs
… dejafu-tests/lib/Examples/AutoUpdate.hs
… dejafu-tests/lib/Examples/ClassLaws.hs
… dejafu-tests/lib/Examples/Logger.hs
… dejafu-tests/lib/Examples/ParMonad.hs
… dejafu-tests/lib/Examples/ParMonad/Direct.hs
… dejafu-tests/lib/Examples/ParMonad/DirectInternal.hs
… dejafu-tests/lib/Examples/Philosophers.hs
… dejafu-tests/lib/Examples/SearchParty.hs
… dejafu-tests/lib/Examples/SearchParty/Impredicative.hs
… dejafu-tests/lib/Integration.hs
… dejafu-tests/lib/Integration/Async.hs
… dejafu-tests/lib/Integration/Litmus.hs
… dejafu-tests/lib/Integration/MultiThreaded.hs
… dejafu-tests/lib/Integration/Refinement.hs
… dejafu-tests/lib/Integration/Regressions.hs
… dejafu-tests/lib/Integration/SingleThreaded.hs
… dejafu-tests/lib/QSemN.hs
… dejafu-tests/lib/Test/Tasty/LeanCheck.hs
… dejafu-tests/lib/Unit.hs
… dejafu-tests/lib/Unit/Properties.hs
… dejafu/CHANGELOG.rst
… dejafu/Test/DejaFu/Conc/Internal.hs