## Summary

Gives dejafu-tests some much-needed love.

While this mostly touches dejafu-tests, it does add a `Test.DejaFu.SCT.Internal.DPOR.independent` function, which will be necessary to implement #183.

- [x] Improve code quality
    - [x] Turn on normal style & lint checks in CI
    - [x] Include perf output in CI
- [x] Improve test coverage
    - [x] Add Par monad case study to dejafu-tests examples (#184)
    - [x] Dependency function testing idea (https://github.com/barrucadu/dejafu/issues/183#issuecomment-363871314)
    - [x] Add tests for getNumCapabilities / setNumCapabilities with multiple threads (#96)
    - [x] Add a test for liftIO (#93)
    - [x] Re-run the code coverage and identify new omissions
- [x] Shuffle around test hierarchy so it's clear why things are where they are and what they do
- [x] Migrate to Hedgehog for all property tests

**Put aside form later:** Was going to also solve these issues, but I'm taking them off the list as I'm not actually sure how best to test these:

- Add tests for named threads, vars, and refs (#98)
- Add tests for resultsSet, resultsSet', resultsSetDiscard, and resultsSetDiscard' (#97)
- Add a test for DiscardTrace (#95)

**Correctness concerns:** One of the changes I made is to move the LeanCheck tests over to Hedgehog.  Given that I found QuickCheck totally useless for those tests, it's worth justifying why I think the Hedgehog tests are at least as good as the LeanCheck tests were.

1. I severely constrained the range of IDs generated, which should increase dependencies.  In hindsight this is something I should have done with QuickCheck too.
2. I seeded some faults which LeanCheck found (051d810, 5bc3eb2, 03ca3c9), and Hedgehog also found them.
3. It found a new bug (#204, https://github.com/rudymatela/leancheck/issues/4)
4. Given that the LeanCheck tests used `(==>)`, we have no idea how many of those tests vacuously passed anyway.

**Related issues:** There isn't one mega-issue for this, it's just a resolution of all the smaller ones mentioned above.

I fixed some other issues which I found while working:

- forkOS is recorded in trace as if it were a fork (#205)
- dependent' is missing a case for BlockedSTM (#204)
- Assignment of thread IDs to commit threads should be more predictable. (#193)
- updateMaskState is missing a case for a thread stopping (#192)
- Incorrect getNumCapabilities / setNumCapabilities dependency (#189)

So there should be a patch release of dejafu after merging this.

## Benchmark results 

While not exactly a performance issue, there are now a bunch more tests and some core components have been changed (`HUnit` / `test-framework` ---> `tasty`, `leancheck` / `QuickCheck` ---> `hedgehog`).

**Before:** (275 tests)

```
$ stack exec -- dejafu-tests +RTS -s

  86,610,404,040 bytes allocated in the heap
  13,030,206,280 bytes copied during GC
      13,360,424 bytes maximum residency (4749 sample(s))
          99,112 bytes maximum slop
              37 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     78806 colls,     0 par   29.948s  29.950s     0.0004s    0.0349s
  Gen  1      4749 colls,     0 par    2.546s   2.548s     0.0005s    0.0018s

  TASKS: 173937 (173933 bound, 4 peak workers (4 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  121.818s  (117.132s elapsed)
  GC      time   32.494s  ( 32.497s elapsed)
  EXIT    time    0.001s  (  0.011s elapsed)
  Total   time  154.314s  (149.640s elapsed)

  Alloc rate    710,980,444 bytes per MUT second

  Productivity  78.9% of total user, 78.3% of total elapsed
```

- Time / test: 0.6s
- Residency / test: 50kB

**After:** (377 tests)

```
$ stack exec -- dejafu-tests +RTS -s

 197,509,567,664 bytes allocated in the heap                                                                                                             
  52,702,831,984 bytes copied during GC
      18,537,592 bytes maximum residency (17340 sample(s))
         222,088 bytes maximum slop
              38 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     172816 colls,     0 par   102.563s  102.575s     0.0006s    0.0470s
  Gen  1     17340 colls,     0 par   10.591s  10.590s     0.0006s    0.0021s

  TASKS: 225347 (225342 bound, 5 peak workers (5 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.000s elapsed)
  MUT     time  208.284s  (200.647s elapsed)
  GC      time  113.154s  (113.165s elapsed)
  EXIT    time    0.002s  (  0.008s elapsed)
  Total   time  321.441s  (313.820s elapsed)

  Alloc rate    948,271,291 bytes per MUT second

  Productivity  64.8% of total user, 63.9% of total elapsed
```

- Time / test: 0.9s
- Residency / test: 50kB

So the tests are now, on average, a little slower.  There are two clear hot spots: the Par monad example, and the SCT property tests.

I suspect that the Hedgehog generators for `ThreadAction` and the like could be improved.  Or maybe a new generator written which is more likely to generate dependent actions.  **This is the same problem I had with QuickCheck but, then again, we don't know if LeanCheck was actually doing any better.**  But this is tricky, as I don't want to accidentally bake in assumptions of implementation details which may change.